arguments matching
git-svn-id: https://asn1c.svn.sourceforge.net/svnroot/asn1c/trunk@604 59561ff5-6e30-0410-9f3c-9617f08c8826

+++ asn1c/enber.c
@@ -328,8 +328,8 @@ process_line(const char *fname, char *line, int lineno) {
 	}
 	if(ret != tl_len) {
 		fprintf(stderr, "%s: Cannot encode TL at line %d "
-			"in the given number of bytes (%d!=%d)\n",
-			fname, lineno, ret, tl_len);
+			"in the given number of bytes (%ld!=%ld)\n",
+			fname, lineno, (long)ret, (long)tl_len);
 		exit(EX_DATAERR);
 	}
 	if(constr) *buf |= 0x20;	/* Enable "constructed" bit */
